Refactoring code
Add React Router
